% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_database_aa3.R
\name{build_db_aa3}
\alias{build_db_aa3}
\title{Build database}
\usage{
build_db_aa3(aa3_combine)
}
\arguments{
\item{aa3_combine}{aa3 Data}
}
\value{
2 data frame constitués des données de calibration dans le premier et
des données pour les échantillons dans le second.
}
\description{
Build database
}
\examples{
# convert_aa3("inst/extra_data/181018E.TXT",
# "inst/extra_data/181018E.xlsx") -> aa3_combine
# plot_aa3(aa3_combine)
# new_calb_aa3(aa3_combine, filter = list(Ptot = c(50,25), Ntot = c(0.1, 0.5)))
# build_db_aa3(aa3_combine)

}
