% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calb_aa3.R
\name{calb.aa3}
\alias{calb.aa3}
\title{calb_aa3 function}
\usage{
calb.aa3(obj, filter_list = NULL)
}
\arguments{
\item{filter_list}{liste de concentration a supprimer pour recalculer
la regression. ( ex : list(Ptot = c(1,2)) -> filter )}

\item{aa3_combine}{aa3 Data}
}
\value{
un dataframe
}
\description{
calb_aa3 function
}
\examples{
# todo

}
