% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calb_aa3.R
\name{calb_aa3}
\alias{calb_aa3}
\title{calb_aa3}
\usage{
calb_aa3(x)
}
\arguments{
\item{x}{aa3 data}
}
\value{
Une liste contenant un dataframe avec les donnees CALB,
les graphes de controle pour la calibration, un dataframe avec les parametres
des regressions et un dataframe avec les donnees SAMP filtrees
}
\description{
calb_aa3
}
\examples{
# TODO

}
