% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calb_correction_aa3.R
\name{calb_correction_aa3}
\alias{calb_correction_aa3}
\title{calb_correction}
\usage{
calb_correction_aa3(aa3, filter_list)
}
\arguments{
\item{aa3}{calb_aa3 Data}

\item{filter_list}{liste de concentration a supprimer pour recalculer
la régression. (ex : list(Ptot = c(1,2)) -> filter )}
}
\value{
Une liste contenant un dataframe avec les donnees CALB, les graphes
de controle pour la calibration et la ou les nouvelle(s) droite(s)
de calibration, un dataframe avec les parametres des regressions et
un dataframe avec les donnees SAMP recalculee.
}
\description{
calb_correction
}
\examples{
# TO DO

}
