% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aa3.R
\name{convert_aa3}
\alias{convert_aa3}
\title{Conversion of aa3 file txt and excell}
\usage{
convert_aa3(file_aa3_txt, file_aa3_xlsx)
}
\arguments{
\item{file_aa3_txt}{txt file}

\item{file_aa3_xlsx}{xlsx file}
}
\value{
un objet converti
}
\description{
Conversion du fichier fourni par l autoanalysuer seal AA3 au format txt  a l aide d un fichier complementaire au format xlsx
}
\examples{

#TODO





}
