% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth}
\alias{growth}
\title{Buoyant weight of corals over time}
\format{
An object of class \code{data.frame} with 32 rows and 7 columns.
}
\usage{
growth
}
\description{
blablabla
}
\examples{
head(growth)

}
\keyword{datasets}
