% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sampdb_aa3.R
\name{merge_sampdb_aa3}
\alias{merge_sampdb_aa3}
\title{merge_sampdb_aa3}
\usage{
merge_sampdb_aa3(sampdb_org, sampdb_inorg)
}
\arguments{
\item{sampdb_org}{calb_aa3$sampdb for organic}

\item{sampdb_inorg}{calb_aa3$sampdb for inorganic}
}
\value{
un dataframe avec les donnees combinées des analyses organiques et
inorganique
}
\description{
merge_sampdb_aa3
}
\examples{
# TO DO

}
