% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_calb_aa3.R
\name{new_calb_aa3}
\alias{new_calb_aa3}
\title{calb_aa3 function}
\usage{
new_calb_aa3(aa3_combine, filter_list)
}
\arguments{
\item{aa3_combine}{aa3 Data}

\item{filter_list}{liste de concentration a supprimer pour recalculer
la regression. ( ex : list(Ptot = c(1,2)) -> filter )}
}
\value{
un dataframe
}
\description{
calb_aa3 function
}
\examples{
# aa3_combine <- convert_aa3("inst/extra_data/181018E.TXT", "inst/extra_data/181018E.xlsx")
# filter_list <- list(Ptot =  c(50, 25), Ntot = c(0.1, 0.5, 1, 2))
# plot_aa3(aa3_combine)
# new_calb_aa3(aa3_combine, filter_list)

}
