% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa3.R
\name{plot.aa3}
\alias{plot.aa3}
\title{plot aa3 file}
\usage{
\method{plot}{aa3}(obj, graph_type = "NO", old_data = FALSE)
}
\arguments{
\item{obj}{aa3 object}

\item{graph_type}{definir le type de graphique que l'on souhaite sortir, par
defaut "NO". "ALL" pour sortir le graphique avec toute les donnees, "lm" pour
sortir le graphique de la régression pour la calibration}

\item{old_data}{logical value to use data before modification, by default
FALSE}
}
\value{
Une liste contenant un dataframe avec les donnees CALB,
les graphes de controle pour la calibration, un dataframe avec les parametres
des regressions et un dataframe avec les donnees SAMP filtrees
}
\description{
plot aa3 file
}
\examples{
#TODO

}
