% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa3.R
\name{plot_aa3}
\alias{plot_aa3}
\title{plot aa3 file}
\usage{
plot_aa3(aa3_combine)
}
\arguments{
\item{aa3_combine}{aa3 data}
}
\value{
Une liste contenant un dataframe avec les donnees CALB,
les graphes de controle pour la calibration, un dataframe avec les parametres
des regressions et un dataframe avec les donnees SAMP filtrees
}
\description{
plot aa3 file
}
\examples{
#TODO

}
