% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp_inorga}
\alias{samp_inorga}
\title{Sample inorganic with seal aa3}
\format{
An object of class \code{data.frame} with 6 rows and 20 columns.
}
\usage{
samp_inorga
}
\description{
blablabla
}
\examples{
head(samp_inorga)

}
\keyword{datasets}
