% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samp_orga}
\alias{samp_orga}
\title{Sample organic with seal aa3}
\format{
An object of class \code{data.frame} with 6 rows and 20 columns.
}
\usage{
samp_orga
}
\description{
blablabla
}
\examples{
head(samp_orga)

}
\keyword{datasets}
