% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_aa3.R
\name{validate_data_aa3}
\alias{validate_data_aa3}
\title{Validate_data_aa3}
\usage{
validate_data_aa3(aa3_combine, filter_list = NULL)
}
\arguments{
\item{aa3_combine}{aa3 Data}

\item{filter_list}{liste de concentration a supprimer pour recalculer
la régression. ( ex : list(Ptot = c(1,2)) -> filter )}
}
\value{
Une liste contenant un dataframe avec les donnees CALB,
une liste de dataframe avec les paramètres des regressions et
un dataframe avec les donnees SAMP eventuellement recalculee. Affiche
egalement les graphes des nouvelles courbes de calibration.
}
\description{
Validate_data_aa3
}
\examples{
#TODO

}
