% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calb_aa3.R
\name{calb_aa3}
\alias{calb_aa3}
\title{Title}
\usage{
calb_aa3(x)
}
\arguments{
\item{x}{aa3 data}
}
\value{
a list containing a data.frame with the parameters of the regression (intercept, values, r squared and n) and a list of graphs for each method of dosage (plot of all data and linear regression plot)
}
\description{
Title
}
\examples{
# TODO

}
