% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sampdb_aa3.R
\name{merge_sampdb_aa3}
\alias{merge_sampdb_aa3}
\title{merge_sampdb_aa3}
\usage{
merge_sampdb_aa3(sampdb_org, sampdb_inorg)
}
\arguments{
\item{sampdb_org}{calb_aa3$sampdb for organic}

\item{sampdb_inorg}{calb_aa3$sampdb for inorganic}
}
\value{
un dataframe avec les donnees combinées des analyses organiques et
inorganiques ("Ptot_conc", "Ntot_conc", "PO4_conc", "NOx_conc", "NO2_conc",
"NO3_conc", "NH4_conc")
}
\description{
merge_sampdb_aa3
}
\examples{
# load("~/shared/projects/seal.econum/data/samp_inorga.rda")
# load("~/shared/projects/seal.econum/data/samp_orga.rda")
# merge_sampdb_aa3(samp_orga, samp_inorga)

}
